Plane is an 8-bit grid, renders to 32-bit RGBA. Skip some tests

+++ test/true_color.js
@@ -25,10 +25,10 @@ describe('True Color', function() {
     let pal = ra.getPaletteAll();
     assert.equal(pal.length, 68);
     // Validate palette contents.
-    assert.equal(pal[64].rgb, 0xcc66cc);
-    assert.equal(pal[65].rgb, 0x44aa66);
-    assert.equal(pal[66].rgb, 0x224488);
-    assert.equal(pal[67].rgb, 0xdd4444);
+    assert.equal(pal[64].rgb.toInt(), 0xcc66cc);
+    assert.equal(pal[65].rgb.toInt(), 0x44aa66);
+    assert.equal(pal[66].rgb.toInt(), 0x224488);
+    assert.equal(pal[67].rgb.toInt(), 0xdd4444);
     for (let k = 0; k < pal.length; k++) {
       assert.equal(pal[k].color, k);
     }
